refactor: port over "addClassesToSlottedChildren" method to BoltElement as a decorator; move existing decorators to now live in standalone files
