+++ packages/base-element/src/lib/decorators/index.js
@@ -1,0 +1,6 @@
+export { jsonSchemaProps } from './json-schema-props';
+export { slottedChildrenDecorator } from './add-classes-to-slotted-children';
+export { conditionalShadowDom } from './conditional-shadow-dom';
+export { convertInitialTags } from './convert-initial-tags';
+export { lazyStyles } from './lazy-styles';
+export { renderAndRenderedEvents } from './render-and-rendered-events';
